Some changes

+++ server/svr.py
@@ -3,10 +3,8 @@ import os
 from keras.preprocessing import image
 import numpy as np
 import tensorflow as tf
-from flask_cors import CORS
 
 app = Flask(__name__)
-CORS(app)
 
 @app.route("/")
 def home():
@@ -17,7 +15,7 @@ model = tf.keras.models.load_model("./model/modelo_treinado.keras")
 img_height = 180
 img_width = 180
 
-target_img = os.path.join(os.getcwd(), 'images')
+target_img = os.path.join(os.getcwd(), 'server', 'images')
 
 if not os.path.exists(target_img):
     os.makedirs(target_img)
@@ -36,38 +34,27 @@ def read_image(filename):
     return x
 
 @app.route('/predict', methods=['POST'])
-@app.route('/predict', methods=['POST'])
 def predict():
-  if request.method == 'POST':
-    try:
-      # Print entire request object
-      print(f"Request headers: {request.headers}")
-      print(f"Request data: {request.data}")
-      print(f"Request files: {request.files}")
-      file = request.files['file']
-      #if file and allowed_file(file.filename):
-      filename = file.filename
-      file_path = os.path.join(target_img, filename)
-      file.save(file_path)
-      img = read_image(file_path)
-      predictions = model.predict(img)
-      class_index = np.argmax(predictions, axis=1)[0]
-      class_names = ['leuko', 'normal']  # Substitua pelos seus nomes de classe
-      predicted_class = class_names[class_index]
-      confidence = predictions[0][class_index]
-      os.remove(file_path)
-      return jsonify({
-        "predicted_class": predicted_class,
-        "confidence": float(confidence),
-        "user_image": file_path
-      })
-      #else:
-        #return jsonify({"error": "Unable to read the file. Please check file extension"})
-    except Exception as e:
-      # Catch any other unexpected errors
-      return jsonify({"error": f"An error occurred: {str(e)}"})
-  else:
-    return jsonify({"error": "Invalid request method. Please use POST."})
+    if request.method == 'POST':
+        file = request.files['image']
+        if file and allowed_file(file.filename):
+            filename = file.filename
+            file_path = os.path.join(target_img, filename)
+            file.save(file_path)
+            img = read_image(file_path)
+            predictions = model.predict(img)
+            class_index = np.argmax(predictions, axis=1)[0]
+            class_names = ['leuko', 'normal']  # Substitua pelos seus nomes de classe
+            predicted_class = class_names[class_index]
+            confidence = predictions[0][class_index]
+            os.remove(file_path)
+            return jsonify({
+                "predicted_class": predicted_class,
+                "confidence": float(confidence),
+                "user_image": file_path
+            })
+        else:
+            return jsonify({"error": "Unable to read the file. Please check file extension"})
 
-if __name__ == '_main_':
+if __name__ == '__main__':
     app.run(debug=True, use_reloader=False, port=8000)
